[Patchset 1] Add Android emulator support
Add prebuilts and scripts for the Android emulator
and all the required SDK harness.
Also introduces a run_android_emulator script that
sets up the necessary .avd files and starts the
emulator using the configuration files checked in
into //build/android_emulators.
In the next CLs this will be wired up to tests.

Change-Id: I56b7cecaaea48ab5f28682c2f341f443c04b134b