[Patchset 4] Introduce protozero protoc compiler plugin (3)
Introduce the plugin can be passed to protoc --plugin to
generate ProtoZero headers from .proto files.
Also introduce the build system incantations required to
generate protobufs (with and without the protozero plugin).
Finally add a end-to-end test to check that the protos
written by ProtoZero are readable by the official protobuf
library.

Change-Id: I1cd4dde5dc7a3efee2d7de527e31195cfdff4fc9