[Patchset 1] Travis CI: switch Trusty bot to use the Ubuntu version of clang
When setting compiler: clang Travis pull a newer version of clang than
the version that is shipped with Trusty. This is good for all the other
configs, but we want to keep at least one target in the matrix to match
the default version of clang that we have on our workstations.

Change-Id: I67cc365343eb8d048ac326f343c81a6d31a6bec2