[Patchset 4] [IPC 3] Add Deferred class to handle async IPC responses.
Deferred<T> is a wrapper for handling async results.
The problem this is solving is the following: For each result argument of the
methods generated from the .proto file:
- The client wants to see something on which it can Bind() a callback, which
  is invoked asynchrnously once reply is received fromt he host.
- The host wants to expose something to embedder that implements the IPC
  methods to allow them to provide an asynchronous reply back to the client.
  Eventually even more than once, for the case streaming replies.

In both cases we want to make sure that callbacks don't get lost along the
way. To address this, this class will automatically reject the callbacks
if they are not resolved at destructor time (or the object is std::move()'d).

The client is supposed to use this class as follows:
 class GreeterProxy {
    void SayHello(const HelloRequest&, Deferred<HelloReply> reply)
 }
...
Deferred<HelloReply> reply;
reply.Bind([] (AsyncResult<HelloReply> reply) {
  std::cout << reply.success() ? reply->message : "failure";
});
host_proxy_instance.SayHello(req, std::move(reply));

The host instead is supposed to use this as follows:
 class GreeterImpl : public Greeter {
   void SayHello(const HelloRequest& req, Deferred<HelloReply> reply) {
      AsyncResult<HelloReply> reply = AsyncResult<HelloReply>::New();
      reply->set_greeting("Hello " + req.name)
      reply.Resolve(std::move(reply));
   }
 }
Or for more complex cases, the deferred object can be std::move()'d outside
and the reply can continue asynchrnously later.

Bug: 68854111
Test: ipc_unittests --gtest_filter=DeferredTest.*
Change-Id: Ibc2b11e92e3bb2d29db9f3509b0eeec556697ab0