[Patchset 4] Add method to create FtraceCpuReader::Config
Callers want to think it terms of names of events, but the parsing
code needs to think in terms of ftrace event ids.

FtraceCpuReader::CreateConfig take a set of names, looks up the
corresponding ftrace event id for each name and sets that id to true
in a bit vector so we can quickly test whether it is enabled later.

Change-Id: Icc6a013644d4c9b9844af6d995683f53838d3db5