[Patchset 5] Use hermetic clang with sanitizers
While so far we had no problems building the release/debug configs with
clang back to 3.4, turns out the situation is more difficult with
sanitizers. asan is broken with-3.4 and msan is broken with clang-3.9.
This makes harder for devs to reproduce an asan failure as seen on the
CI bots.
This CL introduces the ability to use a hermetic compiler, pinning the
same clang tarball that Chrome uses (% keeping it up to date) and by
default enables the hermetic version of clang when using any of the
sanitizers (overridable by setting is_hermetic_clang=true).
This still leaves the default debug/release configurations built with
the system clang, so we keep coverage for compilers available on
Ubuntu Trusty on the CI.

Change-Id: I824b729b6ae76b4b17dfacf0c44982414ba6ceda