[Patchset 3] [IPC 6] Add code geneator
Introduces a plugin for protoc to generate .ipc.{cc,h} files for
.proto services. Also introduces the build file library to integrate
the plugin into the build system.
It has some rough edges (see TODOs, e.g., will fail if the types
of the reply/request aren't defined in the same proto file), but
should be good to go for a v0. At the end of the day any sort of
bug in this code will just produce a build failure.

Bug: 68854111
Change-Id: I2b9a21af972b6ae5350dffce2d1838e560a967c9